requirements.txt
google-cloud-storage==1.14.0
grpcio
Flask==1.0.2
google-cloud-vision==1.0.0